Refactor Database

+++ app/src/main/java/al/bruno/financaime/data/source/local/dao/CategoriesDao.kt
@@ -9,9 +9,9 @@ interface CategoriesDao {
     @Insert
     fun insert(categories: Categories) : Single<Long>
     @Update
-    fun update(categories: Categories) : Single<Long>
+    fun update(categories: Categories) : Single<Int>
     @Delete
-    fun delete(categories: Categories) : Single<Long>
+    fun delete(categories: Categories) : Single<Int>
     @Query("SELECT * FROM categories")
     fun categories() : Single<List<Categories>>
 }
